+++ DataDefinitions/Faction.cs
@@ -36,9 +36,6 @@ namespace EddiDataDefinitions
         // Not intended to be user facing
 
         /// <summary> The faction's EDSM ID </summary>
-        public long? EDDBID { get; set; }
-
-        /// <summary> The faction's EDSM ID </summary>
         public long? EDSMID { get; set; }
 
         /// <summary> The faction's allegiance </summary>