+++ Tests/DataDefinitionTests.cs
@@ -20,7 +20,6 @@ namespace UnitTests
         {
             CommodityDefinition commodity = CommodityDefinition.FromEDName("$ReactiveArmour_Name;");
             Assert.AreEqual("Reactive Armour", commodity.invariantName);
-            Assert.IsNotNull(commodity.EDDBID);
         }
 
         [TestMethod]
@@ -28,7 +27,6 @@ namespace UnitTests
         {
             CommodityDefinition commodity = CommodityDefinition.FromEDName("$MagicStuff_Name;");
             Assert.AreEqual("magicstuff", commodity.invariantName);
-            Assert.IsNull(commodity.EDDBID);
         }
 
         [TestMethod]